fix for multiple tab scenario
if the same page is being opened then two tab containers will exist. One will be in the process of being destroyed and the other will be in the process of being created. Handle this in the widget by having the dojo query return a list and do a check on each item for the _beingDestroyed flag.